requirements.txt
aioamqp==0.14.0
aioredis==1.2.0
asgiref==3.2.2
asn1crypto==0.24.0
astroid==2.4.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
autopep8==1.4.3
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
backports.csv==1.0.7
certifi==2020.4.5.1
cffi==1.12.3
channels==2.2.0
channels-rabbitmq==1.2.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1
django-background-tasks==1.2.0
django-compat==1.0.15
django-crispy-forms==1.9.0
django-import-export==1.2.0
django-mathfilters==1.0.0
django-pandas==0.6.1
django-pyodbc-azure==2.1.0.0
django-storages==1.9.1
djangorestframework==3.10.3
djangorestframework-datatables==0.5.0
et-xmlfile==1.0.1
gitdb2==2.0.5
GitPython==3.0.2
gunicorn==19.9.0
h11==0.8.1
hiredis==1.0.0
httptools==0.0.13
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
jdcal==1.4.1
joblib==0.13.2
lazy-object-proxy==1.4.2
mccabe==0.6.1
msgpack==0.6.1
numpy==1.17.2
odfpy==1.4.0
openpyxl==2.6.3
pamqp==2.3.0
pandas==0.25.1
Pillow==7.1.1
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.5.0
pylint-django==2.0.6
pylint-plugin-utils==0.5
pyodbc==4.0.27
PyPDF2==1.26.0
python-dateutil==2.8.0
python-memcached==1.59
pytz==2019.2
PyYAML==5.1.2
requests==2.23.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
smmap2==2.0.5
sqlparse==0.3.0
tablib==0.13.0
toml==0.10.0
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.9
uvicorn==0.9.0
uvloop==0.13.0
websockets==8.0.2
whitenoise==4.1.3
wrapt==1.11.2
wsproto==0.15.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.1.0
